+++ CHANGES.rst
@@ -5,6 +5,7 @@ unreleased
 - Renamed OdeSys to ODESys (OdeSys left as a deprecated alias)
 - New arguments to ODESys: dep_by_name, par_by_name, param_names, latex_names, latex_param_names
 - New kwargs: first_step_{cb,expr,factory} in ODESys, SymbolicSys & SymbolicSys.from_callback respectively.
+- SymbolicSys.jacobian_singular() returns bool (uses cse and LUdecomposition raising ValueError)
 
 v0.7.0
 ======